+++ Runtime/TESTMONO.cs
@@ -1,0 +1,25 @@
+using UnityEngine;
+
+
+namespace SimpleMan
+{
+    public class TESTMONO : MonoBehaviour
+    {
+        //------FIELDS
+        public string rar;
+
+
+
+        //------PROPERTIES
+
+
+
+
+        //------EVENTS
+
+
+
+
+        //------METHODS
+    }
+}